requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.6.0
async-timeout==4.0.3
attrs==22.2.0
build==0.10.0
CacheControl==0.12.11
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
crashtest==0.4.1
cryptography==39.0.2
distlib==0.3.8
dulwich==0.21.3
filelock==3.15.1
frozenlist==1.4.1
greenlet==3.0.3
html5lib==1.1
idna==3.4
importlib-metadata==6.0.0
installer==0.6.0
jaraco.classes==3.2.3
jeepney==0.8.0
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
magic-filter==1.0.12
more-itertools==9.1.0
msgpack==1.0.5
multidict==6.0.5
packaging==23.0
pexpect==4.8.0
pipenv==2023.2.18
pkginfo==1.9.6
platformdirs==2.6.2
poetry==1.4.0
poetry-core==1.5.1
poetry-plugin-export==1.3.0
ptyprocess==0.7.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pyproject_hooks==1.1.0
pyrsistent==0.19.3
python-dotenv==1.0.1
rapidfuzz==2.13.7
requests==2.28.2
requests-toolbelt==0.10.1
SecretStorage==3.3.3
shellingham==1.5.0.post1
six==1.16.0
SQLAlchemy==2.0.29
tomli==2.0.1
tomlkit==0.11.6
trove-classifiers==2023.3.9
typing_extensions==4.12.2
urllib3==1.26.15
virtualenv==20.20.0
virtualenv-clone==0.5.7
webencodings==0.5.1
xattr==0.10.1
yadisk==2.1.0
yarl==1.9.4
zipp==3.15.0
